MiniProject/requirements.txt
pip
Flask
requests
requests_cache
cassandra_driver